Move multi-device serde functions from _multi_device into serde
Relocate the proto <-> dataclass conversion functions
(serialize/deserialize_model_configuration,
serialize/deserialize_node_device_configuration and their private
helpers) from _multi_device.py to serde.py, placed in the deserialize
and serialize sections alongside the existing glue. _multi_device.py now
holds only the object-bound dataclasses and the private checker, and no
longer imports onnx — a cleaner data-model module. serde.py exports the
four public converters in __all__. Update test references to call them
via serde.
