+++ src/onnx_ir/_multi_device.py
@@ -19,25 +19,14 @@ __all__ = [
     "ShardedDim",
     "ShardingSpec",
     "SimpleShardedDim",
-    "deserialize_model_configuration",
-    "deserialize_node_device_configuration",
-    "serialize_model_configuration",
-    "serialize_node_device_configuration",
 ]
 
 import dataclasses
-import logging
-from collections.abc import Mapping
 from typing import TYPE_CHECKING
-
-import onnx  # noqa: TID251
-
-from onnx_ir._core import SymbolicDim, Value
 
 if TYPE_CHECKING:
     from onnx_ir import _core
-
-logger = logging.getLogger(__name__)
+    from onnx_ir._core import SymbolicDim, Value
 
 
 @dataclasses.dataclass(frozen=True)
@@ -101,177 +90,6 @@ class NodeDeviceConfiguration:
     configuration: ModelConfiguration | None = None
     sharding_spec: tuple[ShardingSpec, ...] = ()
     pipeline_stage: int | None = None
-
-
-def deserialize_model_configuration(
-    proto: onnx.DeviceConfigurationProto,
-) -> ModelConfiguration:
-    return ModelConfiguration(
-        name=proto.name,
-        num_devices=proto.num_devices,
-        device=tuple(proto.device),
-    )
-
-
-def serialize_model_configuration(
-    configuration: ModelConfiguration,
-) -> onnx.DeviceConfigurationProto:
-    proto = onnx.DeviceConfigurationProto()
-    proto.name = configuration.name
-    proto.num_devices = configuration.num_devices
-    proto.device.extend(configuration.device)
-    return proto
-
-
-def _deserialize_simple_sharded_dim(proto: onnx.SimpleShardedDimProto) -> SimpleShardedDim:
-    if proto.HasField("dim_value"):
-        dim: int | SymbolicDim | None = proto.dim_value
-    elif proto.HasField("dim_param"):
-        dim = SymbolicDim(proto.dim_param)
-    else:
-        dim = None
-    return SimpleShardedDim(
-        dim=dim,
-        num_shards=proto.num_shards,
-    )
-
-
-def _serialize_simple_sharded_dim(
-    simple_sharding: SimpleShardedDim,
-) -> onnx.SimpleShardedDimProto:
-    proto = onnx.SimpleShardedDimProto()
-    if isinstance(simple_sharding.dim, int):
-        proto.dim_value = simple_sharding.dim
-    elif (
-        isinstance(simple_sharding.dim, SymbolicDim) and simple_sharding.dim.value is not None
-    ):
-        proto.dim_param = simple_sharding.dim.value
-    proto.num_shards = simple_sharding.num_shards
-    return proto
-
-
-def _deserialize_sharded_dim(proto: onnx.ShardedDimProto) -> ShardedDim:
-    return ShardedDim(
-        axis=proto.axis,
-        simple_sharding=tuple(
-            _deserialize_simple_sharded_dim(simple_sharding)
-            for simple_sharding in proto.simple_sharding
-        ),
-    )
-
-
-def _serialize_sharded_dim(sharded_dim: ShardedDim) -> onnx.ShardedDimProto:
-    proto = onnx.ShardedDimProto()
-    proto.axis = sharded_dim.axis
-    for simple_sharding in sharded_dim.simple_sharding:
-        proto.simple_sharding.append(_serialize_simple_sharded_dim(simple_sharding))
-    return proto
-
-
-def _resolve_value(tensor_name: str, values: Mapping[str, _core.Value]) -> _core.Value | None:
-    """Resolve a proto ``tensor_name`` to a :class:`~onnx_ir.Value`.
-
-    When the name is not present in ``values`` a placeholder value carrying the
-    name is created so the reference round-trips losslessly. This mirrors how
-    missing node inputs are handled during deserialization.
-    """
-    if not tensor_name:
-        return None
-    value = values.get(tensor_name)
-    if value is None:
-        logger.warning(
-            "ShardingSpec references tensor '%s' which is not found in the current "
-            "scope. Creating a placeholder value.",
-            tensor_name,
-        )
-        value = Value(name=tensor_name)
-    return value
-
-
-def _deserialize_sharding_spec(
-    proto: onnx.ShardingSpecProto, values: Mapping[str, _core.Value]
-) -> ShardingSpec:
-    return ShardingSpec(
-        value=_resolve_value(proto.tensor_name, values),
-        device=tuple(proto.device),
-        index_to_device_group_map=tuple(
-            IndexToDeviceGroupMapEntry(key=entry.key, value=tuple(entry.value))
-            for entry in proto.index_to_device_group_map
-        ),
-        sharded_dim=tuple(_deserialize_sharded_dim(dim) for dim in proto.sharded_dim),
-    )
-
-
-def _serialize_sharding_spec(sharding_spec: ShardingSpec) -> onnx.ShardingSpecProto:
-    proto = onnx.ShardingSpecProto()
-    if sharding_spec.value is not None:
-        name = sharding_spec.value.name
-        if not name:
-            raise ValueError(
-                "Cannot serialize a ShardingSpec whose value has no name. "
-                f"Value: {sharding_spec.value!r}"
-            )
-        proto.tensor_name = name
-    proto.device.extend(sharding_spec.device)
-    for entry in sharding_spec.index_to_device_group_map:
-        map_entry = proto.index_to_device_group_map.add()
-        map_entry.key = entry.key
-        map_entry.value.extend(entry.value)
-    for sharded_dim in sharding_spec.sharded_dim:
-        proto.sharded_dim.append(_serialize_sharded_dim(sharded_dim))
-    return proto
-
-
-def deserialize_node_device_configuration(
-    proto: onnx.NodeDeviceConfigurationProto,
-    values: Mapping[str, _core.Value] | None = None,
-) -> NodeDeviceConfiguration:
-    """Deserialize a node device configuration.
-
-    Args:
-        proto: The proto to deserialize.
-        values: Mapping from value name to :class:`~onnx_ir.Value` used to
-            resolve ``tensor_name`` references. When ``None`` all references
-            resolve to freshly created placeholder values.
-
-    Note:
-        ``configuration`` is resolved to a *placeholder* :class:`ModelConfiguration`
-        carrying the ``configuration_id``. The placeholder is replaced by the real
-        model configuration object by a post-pass once the model is fully built
-        (see :func:`onnx_ir.serde.deserialize_model`).
-    """
-    if values is None:
-        values = {}
-    configuration: ModelConfiguration | None = None
-    if proto.configuration_id:
-        configuration = ModelConfiguration(name=proto.configuration_id, num_devices=0)
-    pipeline_stage = proto.pipeline_stage if proto.HasField("pipeline_stage") else None
-    return NodeDeviceConfiguration(
-        configuration=configuration,
-        sharding_spec=tuple(
-            _deserialize_sharding_spec(spec, values) for spec in proto.sharding_spec
-        ),
-        pipeline_stage=pipeline_stage,
-    )
-
-
-def serialize_node_device_configuration(
-    node_device_configuration: NodeDeviceConfiguration,
-) -> onnx.NodeDeviceConfigurationProto:
-    proto = onnx.NodeDeviceConfigurationProto()
-    if node_device_configuration.configuration is not None:
-        name = node_device_configuration.configuration.name
-        if not name:
-            raise ValueError(
-                "Cannot serialize a NodeDeviceConfiguration whose configuration has "
-                f"no name. Configuration: {node_device_configuration.configuration!r}"
-            )
-        proto.configuration_id = name
-    for sharding_spec in node_device_configuration.sharding_spec:
-        proto.sharding_spec.append(_serialize_sharding_spec(sharding_spec))
-    if node_device_configuration.pipeline_stage is not None:
-        proto.pipeline_stage = node_device_configuration.pipeline_stage
-    return proto
 
 
 def _check_device_configurations(model: _core.Model) -> list[str]:
